Improved the AtomChainDialog and related code in the PtmDialog. Made sure that the result of the toString method for the atomic composition is the same as in Unimod.

+++ src/main/java/com/compomics/util/experiment/biology/ions/ReporterIon.java
@@ -318,7 +318,7 @@ public class ReporterIon extends Ion {
      * @return a boolean indicating whether masses are equal
      */
     public boolean isSameAs(ReporterIon anotherReporterIon) {
-        return theoreticMass == anotherReporterIon.getTheoreticMass();
+        return theoreticMass.doubleValue() == anotherReporterIon.getTheoreticMass(); // @TODO: compare against the accuracy!
     }
 
     /**
